@@ -14,8 +14,8 @@
     "@angular/animations": "^17.3.6",
     "@angular/cdk": "^13.3.9",
     "@angular/common": "^17.3.6",
-    "@angular/compiler": "^13.4.0",
-    "@angular/core": "^13.4.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
     "@angular/forms": "^17.3.6",
     "@angular/http": "^6.1.0",
     "@angular/material": "^13.3.9",
@@ -29,9 +29,9 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^17.3.6",
-    "@angular/cli": "~13.3.11",
+    "@angular/cli": "^14.2.13",
     "@angular/compiler-cli": "^17.3.6",
-    "@angular/language-service": "^13.4.0",
+    "@angular/language-service": "^14.3.0",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
